refactor: create Footer folder
