+++ wealth-sync/src/components/Footer/Footer.tsx
